updating qt psana builds against
The meta.yaml used to specify qt 4.8.5, because that's what I got working with building before - at some point I let it run against later qt to work with spyder and other things - however we just tried to build a package like PSQt against qt 4.8.7 and it didn't work. We just checked in changes to specify the new location for header files with qt 4.8.7, looks like this works against THAT version of qt.